Work with and without crumbs enabled
